Update to kramdown 0.14.2+
this includes a bugfix[1] to avoid a crash when a heading is specified
with no text.

https://github.com/gettalong/kramdown/commit/1db179c7

Unfortunately another change in kramdown has changed behaviour covered
by one of our tests. If 'numeric' entity output is requested, &amp;
still comes out as &amp; rather than &#38;

I think the kramdown change is an error, and will contact them about it.
However, I don't think this behaviour is required by us (I may be
wrong). I've changed our test so that it passes but still reflects
the original intention of the test.